# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrEnvelope:
    properties:
      error:
        type: string
    type: object
  handler.MsgEnvelope:
    properties:
      message:
        type: string
    type: object
...
  model.CreateUserRq:
    properties:
      login:
        minLength: 5
        type: string
      password:
        minLength: 5
        type: string
      user_name:
        minLength: 5
        type: string
    required:
    - login
    - password
    - user_name
    type: object
  model.LogInRq:
    properties:
      login:
        minLength: 5
        type: string
      password:
        minLength: 5
        type: string
    required:
    - login
    - password
    type: object
...
  model.UpdateUserRq:
    properties:
      login:
        type: string
      user_name:
        type: string
    type: object
  model.User:
    properties:
      id:
        minimum: 1
        type: integer
      is_active:
        type: boolean
      login:
        type: string
      user_name:
        type: string
      version:
        type: integer
    required:
    - id
    type: object
host: localhost:8080
info:
  contact: {}
  description: Onelab project
  title: Onelab API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User login input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.LogInRq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      summary: Login user
      tags:
      - user
  /user:
    post:
      consumes:
      - application/json
      description: Create a new user with the input payload
      parameters:
      - description: User information
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserRq'
      produces:
      - application/json
      responses:
        "201":
          description: user created
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "422":
          description: user already exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      summary: Create a new user
      tags:
      - user
  /user/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: user deleted
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: user is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Delete user by id
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get user by id in query param
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "302":
          description: Found
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: user is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Get user by id
      tags:
      - user
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User information
        in: body
        name: input
        schema:
          $ref: '#/definitions/model.UpdateUserRq'
      produces:
      - application/json
      responses:
        "200":
          description: user updated
          schema:
            $ref: '#/definitions/handler.MsgEnvelope'
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "401":
          description: missing or malformed jwt
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "404":
          description: user is not exist
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "409":
          description: edit conflict
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrEnvelope'
      security:
      - ApiKeyAuth: []
      summary: Update user by id
      tags:
      - user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"